Updated return type

+++ solid-net-sample/MathShapeCalc/CalculateArea.cs
@@ -13,7 +13,7 @@ public class CalculateArea
         return sum;
     }
 
-    public double count_shapes(List<Shape> shapes)
+    public int count_shapes(List<Shape> shapes)
     {
         return shapes.Count;
     }
